Move Inflight actors to ut_helpers
