+++ ydb/core/blobstorage/ut_blobstorage/ya.make
@@ -37,6 +37,7 @@ SRCS(
     snapshots.cpp
     space_check.cpp
     sync.cpp
+    ut_helpers.cpp
 )
 
 PEERDIR(
